+++ modulo2/apir-labenusers/apis-labenusers/src/App.js
@@ -1,5 +1,7 @@
 import React from "react";
 import styled from "styled-components";
+import PageOne from "./Pages/PaginaInicial";
+import PageTwo from "./Pages/PaginaSecundaria";
 
 const Button = styled.button`
 width: 105px;
@@ -7,50 +9,36 @@ height: 20px;
 `
 
 class App extends React.Component {
+
   state = {
-    registro: [
-      {
-        id: "",
-        nome: "",
-        email: ""
-      }
-    ],
-    nomeBase: "",
-    emailBase: ""
+    telaBase: false
   }
 
-
-  updateNome = (e) => {
-    console.log("Nome foi atualizado");
-    this.setState({ nomeBase: e.target.value })
+  primeiraTela = () => {
+    this.setState({ telaBase: false })
   }
 
-  updateEmail = (e) => {
-    console.log("Email foi atualizado");
-    this.setState({ emailBase: e.target.value })
-  }
-
-  criaUsuario = (e) => {
-    console.log("Fui clicado");
-    const novoUsuario = {
-      nome: this.state.nomeBase,
-      email: this.state.emailBase
-    }
-    const novoArrayUsuario = [...this.state.registro, novoUsuario]
-    this.setState({ registro: novoArrayUsuario, nomeBase: "", emailBase: "" })
+  segundaTela = () => {
+    this.setState({ telaBase: true })
   }
 
   render() {
 
+    let pagina = {}
+
+    if (this.state.telaBase) {
+      pagina = <PageTwo trocarTela={this.primeiraTela} />
+    } else {
+      pagina = <PageOne trocarTela={this.segundaTela} />
+    }
+
     return (
 
-      <div>
-        <Button>Trocar de Tela</Button>
-        <input type="text" placeholder="Nome" onChange={this.updateNome} value={this.state.nomeBase} />
-        <input type="email" placeholder="Email" onChange={this.updateEmail} value={this.state.emailBase} />
-        <Button onClick={this.criaUsuario} >Criar Usuário</Button>
-      </div>
-    )
+        <div>
+          {pagina}
+        </div>
+
+      )
   }
 }
